apps/functions/video_generation/requirements.txt
functions-framework>=3.4.0
google-genai>=0.1.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
requests>=2.28.0